+++ Entity/Models/Passangers.cs
@@ -13,11 +13,11 @@ namespace Entity.Models
         public string PhoneNumber { get; set; }
         public string Email { get; set; }
         public string Address { get; set; }
-        public double Latitude { get; set; } 
-        public double Longitude { get; set; }
+        public decimal Latitude { get; set; } 
+        public decimal Longitude { get; set; }
         public string CitizenNumber { get; set; }
-        public double? PickupLatitude { get; set; }
-        public double? PickupLongitude { get; set; }
+        public decimal? PickupLatitude { get; set; }
+        public decimal? PickupLongitude { get; set; }
         public int WalkingDistanceInMeters { get; set; }
         public bool UseAlternatePickup { get; set; } = false;
         public int? RouteId { get; set; }
